+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/impl/AbstractPolyglotImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2025, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2017, 2026, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -848,6 +848,7 @@ public abstract class AbstractPolyglotImpl {
 
         public abstract boolean isInterrupted(Object receiver);
 
+        public abstract Object getCauseImpl(Object receiver);
     }
 
     public abstract static class AbstractStackFrameImpl extends AbstractDispatchClass {