+++ sdk/CHANGELOG.md
@@ -15,6 +15,7 @@ This changelog summarizes major changes between GraalVM SDK versions. The main f
 * GR-71402: Added `Value#hasStaticScope` and `Value#getStaticScope` returning the static scope representing static or class-level members associated with the meta object.
 * GR-71643: Added `Context.Builder#exceptionHandler(Consumer<PolyglotException>)` to customize how polyglot exceptions are handled before they are propagated to the host.
 * GR-69929: Reset a `CallTarget`’s execution profile when its associated nmethod is invalidated by HotSpot due to code cache “cold” flushing heuristics.
+* GR-72991: If a `PolyglotException` wraps an exception that has a cause according to the interop protocol (`InteropLibrary.getExceptionCause`), then calling `getCause()` on that `PolyglotExceptions' will return a `PolyglotException` representing that cause. Note that for `PolyglotException` objects that wrap a host exception that has a cause, this means that `getCause` returns a `PolyglotException` that wraps this cause.
 
 ## Version 25.0.0
 * GR-60636 Truffle now stops compiling when the code cache fills up on HotSpot. A warning is printed when that happens.